dstoyko: fixed comments

+++ src/main/java/com/softserveinc/ita/pageobjects/Cart.java
@@ -10,9 +10,9 @@ public class Cart {
         return new InCartProduct(String.format("(//div[@class='cart-product ng-star-inserted'])[%s]", index));
     }
 
-    public float getTotalPrice() {
+    public int getTotalPrice() {
         $x("//div[@class='modal__content modal__content--locked']").shouldNotBe(visible);
 
-        return Float.parseFloat($x("//div[@class='cart-receipt__sum-price']/span").text());
+        return Integer.parseInt($x("//div[@class='cart-receipt__sum-price']/span").text());
     }
 }
